+++ common/src/main/scala/dx/util/CodecUtils.scala
@@ -4,6 +4,7 @@ import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
 import java.security.MessageDigest
 import java.util.Base64
 import java.util.zip.{GZIPInputStream, GZIPOutputStream}
+import javax.xml.bind.DatatypeConverter
 
 object CodecUtils {
   // From: https://gist.github.com/owainlewis/1e7d1e68a6818ee4d50e
@@ -38,5 +39,10 @@ object CodecUtils {
   def md5Checksum(s: String): String = {
     val digest = MessageDigest.getInstance("MD5").digest(s.getBytes)
     digest.map("%02X" format _).mkString
+  }
+
+  def sha1Digest(s: String): String = {
+    val md = MessageDigest.getInstance("SHA-1")
+    DatatypeConverter.printHexBinary(md.digest(s.getBytes))
   }
 }