+++ protocols/src/main/scala/dx/util/protocols/DxFileAccessProtocol.scala
@@ -2,7 +2,7 @@ package dx.util.protocols
 
 import java.nio.charset.Charset
 import java.nio.file.{Files, Path, Paths}
-import dx.api.{DxApi, DxFile, DxFileDescCache, DxFindDataObjects, DxPath, DxProject}
+import dx.api.{DxApi, DxFile, DxFileDescCache, DxFindDataObjects, DxPath, DxProject, Field}
 import dx.util.{
   AbstractAddressableFileNode,
   AddressableFileSource,
@@ -12,23 +12,40 @@ import dx.util.{
   SysUtils
 }
 
+import java.net.URI
+
 case class DxFileSource(dxFile: DxFile, override val encoding: Charset)(
     override val address: String,
-    dxApi: DxApi
+    protocol: DxFileAccessProtocol
 ) extends AbstractAddressableFileNode(address, encoding) {
 
   override def name: String = dxFile.getName
 
   override def folder: String = dxFile.describe().folder
 
+  def project: DxProject = {
+    dxFile.project
+      .getOrElse(protocol.dxApi.project(dxFile.describe(Set(Field.Project)).project))
+  }
+
+  override def getParent: Option[DxFolderSource] = {
+    Some(DxFolderSource(project, folder)(protocol))
+  }
+
+  override def resolve(path: String): AddressableFileSource = {
+    getParent.get.resolve(path)
+  }
+
+  override def uri: URI = URI.create(dxFile.asUri)
+
   override lazy val size: Long = dxFile.describe().size
 
   override def readBytes: Array[Byte] = {
-    dxApi.downloadBytes(dxFile)
+    protocol.dxApi.downloadBytes(dxFile)
   }
 
   override protected def localizeTo(file: Path): Unit = {
-    dxApi.downloadFile(file, dxFile)
+    protocol.dxApi.downloadFile(file, dxFile)
   }
 }
 
@@ -43,6 +60,12 @@ case class DxArchiveFolderSource(dxFileSource: DxFileSource) extends Addressable
 
   override def folder: String = dxFileSource.folder
 
+  override def getParent: Option[AddressableFileSource] = dxFileSource.getParent
+
+  override def resolve(path: String): AddressableFileSource = dxFileSource.resolve(path)
+
+  override def uri: URI = dxFileSource.uri
+
   override protected def localizeTo(dir: Path): Unit = {
     val tempfile = Files.createTempFile("temp", dxFileSource.name)
     val tarOpts = if (dxFileSource.name.endsWith("gz")) "-xz" else "-x"
@@ -56,7 +79,7 @@ case class DxArchiveFolderSource(dxFileSource: DxFileSource) extends Addressable
 }
 
 case class DxFolderSource(dxProject: DxProject, folder: String)(
-    dxApi: DxApi
+    protocol: DxFileAccessProtocol
 ) extends AddressableFileSource {
   override def address: String = s"dx://${dxProject.id}:${folder}"
 
@@ -64,9 +87,31 @@ case class DxFolderSource(dxProject: DxProject, folder: String)(
 
   override def isDirectory: Boolean = true
 
+  override def getParent: Option[DxFolderSource] = {
+    if (folder == "/") {
+      None
+    } else {
+      val parent = Paths.get(folder).getParent.toString match {
+        case p if p.endsWith("/") => p
+        case p                    => s"${p}/"
+      }
+      Some(DxFolderSource(dxProject, parent)(protocol))
+    }
+  }
+
+  override def resolve(path: String): AddressableFileSource = {
+    if (path.endsWith("/")) {
+      DxFolderSource(dxProject, s"${folder}${path}")(protocol)
+    } else {
+      val uri = s"dx://${dxProject.id}:${folder}${path}"
+      protocol.resolve(uri)
+    }
+  }
+
   lazy val listing: Vector[(DxFile, Path)] = {
     val results =
-      DxFindDataObjects(dxApi).apply(Some(dxProject), Some(folder), recurse = true, Some("file"))
+      DxFindDataObjects(protocol.dxApi)
+        .apply(Some(dxProject), Some(folder), recurse = true, Some("file"))
     results.map {
       case (f: DxFile, _) =>
         val relPath = Paths.get(folder).relativize(Paths.get(f.describe().folder))
@@ -79,7 +124,7 @@ case class DxFolderSource(dxProject: DxProject, folder: String)(
     listing.foreach {
       case (dxFile, relPath) =>
         val path = dir.resolve(relPath).resolve(dxFile.getName)
-        dxApi.downloadFile(path, dxFile)
+        protocol.dxApi.downloadFile(path, dxFile)
     }
   }
 }
@@ -122,7 +167,7 @@ case class DxFileAccessProtocol(dxApi: DxApi = DxApi.get,
 
   private def resolveFile(uri: String): DxFileSource = {
     val dxFile = dxFileCache.updateFileFromCache(resolveFileUri(uri))
-    val src = DxFileSource(dxFile, encoding)(uri, dxApi)
+    val src = DxFileSource(dxFile, encoding)(uri, this)
     uriToFileSource += (uri -> src)
     src
   }
@@ -142,18 +187,18 @@ case class DxFileAccessProtocol(dxApi: DxApi = DxApi.get,
       val project = projectName
         .map(dxApi.resolveProject)
         .getOrElse(throw new Exception("project must be specified for a DNAnexus folder URI"))
-      DxFolderSource(project, folder)(dxApi)
+      DxFolderSource(project, folder)(this)
     } else {
       DxArchiveFolderSource(resolveFile(uri))
     }
   }
 
   def resolveNoCache(uri: String): FileSource = {
-    DxFileSource(resolveFileUri(uri), encoding)(uri, dxApi)
+    DxFileSource(resolveFileUri(uri), encoding)(uri, this)
   }
 
   def fromDxFile(dxFile: DxFile): DxFileSource = {
-    DxFileSource(dxFile, encoding)(dxFile.asUri, dxApi)
+    DxFileSource(dxFile, encoding)(dxFile.asUri, this)
   }
 }